[Ecommerce] remove legacy checkout architecture
