+++ bundles/EcommerceFrameworkBundle/CheckoutManager/V7/CommitOrderProcessor.php
@@ -2,26 +2,47 @@
 
 namespace Pimcore\Bundle\EcommerceFrameworkBundle\CheckoutManager\V7;
 
+use Pimcore\Bundle\EcommerceFrameworkBundle\CheckoutManager\CommitOrderProcessorInterface;
 use Pimcore\Bundle\EcommerceFrameworkBundle\Exception\PaymentNotSuccessfulException;
 use Pimcore\Bundle\EcommerceFrameworkBundle\Exception\UnsupportedException;
 use Pimcore\Bundle\EcommerceFrameworkBundle\Model\AbstractOrder;
 use Pimcore\Bundle\EcommerceFrameworkBundle\OrderManager\OrderManagerLocatorInterface;
-use Pimcore\Bundle\EcommerceFrameworkBundle\PaymentManager\Payment\PaymentInterface;
+use Pimcore\Bundle\EcommerceFrameworkBundle\PaymentManager\Status;
 use Pimcore\Bundle\EcommerceFrameworkBundle\PaymentManager\StatusInterface;
+use Pimcore\Bundle\EcommerceFrameworkBundle\PaymentManager\V7\Payment\PaymentInterface;
 use Pimcore\Event\Ecommerce\CommitOrderProcessorEvents;
 use Pimcore\Event\Model\Ecommerce\CommitOrderProcessorEvent;
 use Pimcore\Event\Model\Ecommerce\SendConfirmationMailEvent;
 use Pimcore\Log\ApplicationLogger;
 use Pimcore\Log\FileObject;
+use Pimcore\Logger;
 use Psr\Log\LoggerAwareInterface;
 use Psr\Log\LoggerAwareTrait;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\Lock\LockFactory;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-class CommitOrderProcessor extends \Pimcore\Bundle\EcommerceFrameworkBundle\CheckoutManager\CommitOrderProcessor implements LoggerAwareInterface
+class CommitOrderProcessor implements CommitOrderProcessorInterface, LoggerAwareInterface
 {
     use LoggerAwareTrait;
+
+    const LOCK_KEY = 'ecommerce-framework-commitorder-lock';
+    const LOGGER_NAME = 'commit-order-processor';
+
+    /**
+     * @var OrderManagerLocatorInterface
+     */
+    protected $orderManagers;
+
+    /**
+     * @var LockFactory
+     */
+    private $lockFactory;
+
+    /**
+     * @var string
+     */
+    protected $confirmationMail = '/emails/order-confirmation';
 
     /**
      * @var EventDispatcherInterface
@@ -34,9 +55,15 @@ class CommitOrderProcessor extends \Pimcore\Bundle\EcommerceFrameworkBundle\Chec
     protected $applicationLogger;
 
     /**
-     * @var LockFactory
+     * @var null | string
      */
-    private $lockFactory;
+    protected $lastPaymentStateResponseHash = null;
+
+    /**
+     * @var null | StatusInterface
+     */
+    protected $lastPaymentStatus = null;
+
 
     public function __construct(LockFactory $lockFactory, OrderManagerLocatorInterface $orderManagers, EventDispatcherInterface $eventDispatcher, ApplicationLogger $applicationLogger, array $options = [])
     {
@@ -52,6 +79,68 @@ class CommitOrderProcessor extends \Pimcore\Bundle\EcommerceFrameworkBundle\Chec
         $this->lockFactory = $lockFactory;
     }
 
+    protected function processOptions(array $options)
+    {
+        if (isset($options['confirmation_mail'])) {
+            $this->confirmationMail = $options['confirmation_mail'];
+        }
+    }
+
+    protected function configureOptions(OptionsResolver $resolver)
+    {
+        $resolver->setDefined('confirmation_mail');
+    }
+
+    /**
+     * @param string $confirmationMail
+     */
+    public function setConfirmationMail($confirmationMail)
+    {
+        if (!empty($confirmationMail)) {
+            $this->confirmationMail = $confirmationMail;
+        }
+    }
+
+    /**
+     * @param array|StatusInterface $paymentResponseParams
+     * @param PaymentInterface $paymentProvider
+     *
+     * @return Status|StatusInterface
+     */
+    protected function getPaymentStatus($paymentResponseParams, PaymentInterface $paymentProvider)
+    {
+        $responseHash = md5(serialize($paymentResponseParams));
+
+        if ($this->lastPaymentStateResponseHash === $responseHash) {
+            return $this->lastPaymentStatus;
+        }
+
+        // since handle response can throw exceptions and commitOrderPayment must be executed,
+        // this needs to be in a try-catch block
+        try {
+            $paymentStatus = $paymentProvider->handleResponse($paymentResponseParams);
+        } catch (\Exception $e) {
+            Logger::err($e);
+
+            //create payment status with error message and cancelled payment
+            $paymentStatus = new Status(
+                $paymentResponseParams['orderIdent'],
+                'unknown',
+                'there was an error: ' . $e->getMessage(),
+                StatusInterface::STATUS_CANCELLED
+            );
+        }
+
+        $this->lastPaymentStateResponseHash = $responseHash;
+        $this->lastPaymentStatus = $paymentStatus;
+
+        return $paymentStatus;
+    }
+
+    /**
+     * @inheritDoc
+     * @throws UnsupportedException
+     */
     public function handlePaymentResponseAndCommitOrderPayment($paymentResponseParams, PaymentInterface $paymentProvider)
     {
         $this->logger->info('Payment Provider Response received. ' . print_r($paymentResponseParams, true));
@@ -67,8 +156,42 @@ class CommitOrderProcessor extends \Pimcore\Bundle\EcommerceFrameworkBundle\Chec
         return $this->commitOrderPayment($paymentStatus, $paymentProvider);
     }
 
+
     /**
      * @inheritdoc
+     */
+    public function committedOrderWithSamePaymentExists($paymentResponseParams, PaymentInterface $paymentProvider)
+    {
+        if (!$paymentResponseParams instanceof StatusInterface) {
+            $paymentStatus = $this->getPaymentStatus($paymentResponseParams, $paymentProvider);
+        } else {
+            $paymentStatus = $paymentResponseParams;
+        }
+
+        $order = $this->orderManagers->getOrderManager()->getOrderByPaymentStatus($paymentStatus);
+
+        if ($order && $order->getOrderState() == $order::ORDER_STATE_COMMITTED) {
+            $paymentInformationCollection = $order->getPaymentInfo();
+            if ($paymentInformationCollection) {
+                foreach ($paymentInformationCollection as $paymentInfo) {
+                    if ($paymentInfo->getInternalPaymentId() == $paymentStatus->getInternalPaymentId()) {
+                        if ($paymentInfo->getPaymentState() == $paymentStatus->getStatus()) {
+                            return $order;
+                        } else {
+                            Logger::warning('Payment state of order ' . $order->getId() . ' does not match with new request!');
+                        }
+                    }
+                }
+            }
+        }
+
+        return null;
+    }
+
+    /**
+     * @inheritdoc
+     * @throws UnsupportedException|PaymentNotSuccessfulException
+     * @throws \Exception
      */
     public function commitOrderPayment(StatusInterface $paymentStatus, PaymentInterface $paymentProvider, AbstractOrder $sourceOrder = null)
     {
@@ -138,6 +261,19 @@ class CommitOrderProcessor extends \Pimcore\Bundle\EcommerceFrameworkBundle\Chec
     }
 
     /**
+     * Method for applying additional data to the order object based on payment information
+     * Called in commitOrderPayment() just after updatePayment on OrderAgent is called
+     *
+     * @param AbstractOrder $order
+     * @param StatusInterface $paymentStatus
+     * @param PaymentInterface $paymentProvider
+     */
+    protected function applyAdditionalDataToOrder(AbstractOrder $order, StatusInterface $paymentStatus, PaymentInterface $paymentProvider)
+    {
+        // nothing to do by default
+    }
+
+    /**
      * @inheritdoc
      */
     public function commitOrder(AbstractOrder $order)
@@ -161,6 +297,17 @@ class CommitOrderProcessor extends \Pimcore\Bundle\EcommerceFrameworkBundle\Chec
     }
 
     /**
+     * Implementation-specific processing of order, must be implemented in subclass (e.g. sending order to ERP-system)
+     *
+     * @param AbstractOrder $order
+     */
+    protected function processOrder(AbstractOrder $order)
+    {
+        // nothing to do
+    }
+
+
+    /**
      * @inheritdoc
      */
     protected function sendConfirmationMail(AbstractOrder $order)
@@ -181,6 +328,48 @@ class CommitOrderProcessor extends \Pimcore\Bundle\EcommerceFrameworkBundle\Chec
             } else {
                 $this->logger->error('No Customer found!');
             }
+        }
+    }
+
+    /**
+     * @throws \Exception
+     */
+    public function cleanUpPendingOrders()
+    {
+        $dateTime = new \DateTime();
+        $dateTime->sub(new \DateInterval('PT1H'));
+        $timestamp = $dateTime->getTimestamp();
+
+        $orderManager = $this->orderManagers->getOrderManager();
+
+        //Abort orders with payment pending
+        $list = $orderManager->buildOrderList();
+        $list->addFieldCollection('PaymentInfo');
+        $list->setCondition('orderState = ? AND o_modificationDate < ?', [AbstractOrder::ORDER_STATE_PAYMENT_PENDING, $timestamp]);
+
+        /** @var AbstractOrder $order */
+        foreach ($list as $order) {
+            Logger::warn('Setting order ' . $order->getId() . ' to ' . AbstractOrder::ORDER_STATE_ABORTED);
+            $order->setOrderState(AbstractOrder::ORDER_STATE_ABORTED);
+            $order->save(['versionNote' => 'CommitOrderProcessor::cleanUpPendingOrders - set state to aborted.']);
+        }
+
+        //Abort payments with payment pending
+        $list = $orderManager->buildOrderList();
+        $list->addFieldCollection('PaymentInfo', 'paymentinfo');
+        $list->setCondition('`PaymentInfo~paymentinfo`.paymentState = ? AND `PaymentInfo~paymentinfo`.paymentStart < ?', [AbstractOrder::ORDER_STATE_PAYMENT_PENDING, $timestamp]);
+
+        /** @var AbstractOrder $order */
+        foreach ($list as $order) {
+            $payments = $order->getPaymentInfo();
+
+            foreach ($payments as $payment) {
+                if ($payment->getPaymentState() == AbstractOrder::ORDER_STATE_PAYMENT_PENDING && $payment->getPaymentStart()->getTimestamp() < $timestamp) {
+                    Logger::warn('Setting order ' . $order->getId() . ' payment ' . $payment->getInternalPaymentId() . ' to ' . AbstractOrder::ORDER_STATE_ABORTED);
+                    $payment->setPaymentState(AbstractOrder::ORDER_STATE_ABORTED);
+                }
+            }
+            $order->save(['versionNote' => 'CommitOrderProcessor:cleanupPendingOrders- payment aborted.']);
         }
     }
 }
