+++ bundles/EcommerceFrameworkBundle/OrderManager/OrderManagerLocatorInterface.php
@@ -17,6 +17,8 @@ declare(strict_types=1);
 
 namespace Pimcore\Bundle\EcommerceFrameworkBundle\OrderManager;
 
+use Pimcore\Bundle\EcommerceFrameworkBundle\OrderManager\V7\OrderManagerInterface;
+
 interface OrderManagerLocatorInterface
 {
     public function getOrderManager(string $tenant = null): OrderManagerInterface;
