+++ bundles/EcommerceFrameworkBundle/OrderManager/V7/OrderManagerInterface.php
@@ -17,9 +17,47 @@ namespace Pimcore\Bundle\EcommerceFrameworkBundle\OrderManager\V7;
 use Pimcore\Bundle\EcommerceFrameworkBundle\CartManager\CartInterface;
 use Pimcore\Bundle\EcommerceFrameworkBundle\Exception\UnsupportedException;
 use Pimcore\Bundle\EcommerceFrameworkBundle\Model\AbstractOrder;
+use Pimcore\Bundle\EcommerceFrameworkBundle\OrderManager\OrderAgentInterface;
+use Pimcore\Bundle\EcommerceFrameworkBundle\OrderManager\OrderListInterface;
+use Pimcore\Bundle\EcommerceFrameworkBundle\PaymentManager\StatusInterface;
+use Pimcore\Model\DataObject\Folder;
 
-interface OrderManagerInterface extends \Pimcore\Bundle\EcommerceFrameworkBundle\OrderManager\OrderManagerInterface
+interface OrderManagerInterface
 {
+    /**
+     * @return OrderListInterface
+     */
+    public function createOrderList();
+
+    /**
+     * @param AbstractOrder $order
+     *
+     * @return OrderAgentInterface
+     */
+    public function createOrderAgent(AbstractOrder $order);
+
+    /**
+     * @param int|Folder $orderParentFolder
+     */
+    public function setParentOrderFolder($orderParentFolder);
+
+    /**
+     * @param string $classname
+     */
+    public function setOrderClass($classname);
+
+    /**
+     * @param string $classname
+     */
+    public function setOrderItemClass($classname);
+
+    /**
+     * Looks if order object for given cart already exists, otherwise creates it
+     *
+     * @return AbstractOrder
+     */
+    public function getOrCreateOrderFromCart(CartInterface $cart);
+
     /**
      * @param CartInterface $cart
      *
@@ -33,6 +71,42 @@ interface OrderManagerInterface extends \Pimcore\Bundle\EcommerceFrameworkBundle
      * @return AbstractOrder
      */
     public function recreateOrderBasedOnSourceOrder(AbstractOrder $sourceOrder): AbstractOrder;
+
+    /**
+     * Looks if order object for given cart exists and returns it - it does not create it!
+     *
+     * @param CartInterface $cart
+     *
+     * @return AbstractOrder
+     */
+    public function getOrderFromCart(CartInterface $cart);
+
+    /**
+     * Returns order based on given payment status
+     *
+     * @param StatusInterface $paymentStatus
+     *
+     * @return AbstractOrder
+     */
+    public function getOrderByPaymentStatus(StatusInterface $paymentStatus);
+
+    /**
+     * Builds order listing
+     *
+     * @return \Pimcore\Model\DataObject\Listing\Concrete
+     *
+     * @throws \Exception
+     */
+    public function buildOrderList();
+
+    /**
+     * Build order item listing
+     *
+     * @return \Pimcore\Model\DataObject\Listing\Concrete
+     *
+     * @throws \Exception
+     */
+    public function buildOrderItemList();
 
     /**
      * @param CartInterface $cart
