+++ bundles/EcommerceFrameworkBundle/PaymentManager/PaymentManager.php
@@ -18,7 +18,7 @@ declare(strict_types=1);
 namespace Pimcore\Bundle\EcommerceFrameworkBundle\PaymentManager;
 
 use Pimcore\Bundle\EcommerceFrameworkBundle\PaymentManager\Exception\ProviderNotFoundException;
-use Pimcore\Bundle\EcommerceFrameworkBundle\PaymentManager\Payment\PaymentInterface;
+use Pimcore\Bundle\EcommerceFrameworkBundle\PaymentManager\V7\Payment\PaymentInterface;
 use Psr\Container\ContainerInterface as PsrContainerInterface;
 
 class PaymentManager implements PaymentManagerInterface
