+++ bundles/EcommerceFrameworkBundle/PaymentManager/V7/Payment/PaymentInterface.php
@@ -15,12 +15,19 @@
 namespace Pimcore\Bundle\EcommerceFrameworkBundle\PaymentManager\V7\Payment;
 
 use Pimcore\Bundle\EcommerceFrameworkBundle\OrderManager\OrderAgentInterface;
+use Pimcore\Bundle\EcommerceFrameworkBundle\PaymentManager\StatusInterface;
 use Pimcore\Bundle\EcommerceFrameworkBundle\PaymentManager\V7\Payment\StartPaymentRequest\AbstractRequest;
 use Pimcore\Bundle\EcommerceFrameworkBundle\PaymentManager\V7\Payment\StartPaymentResponse\StartPaymentResponseInterface;
 use Pimcore\Bundle\EcommerceFrameworkBundle\PriceSystem\PriceInterface;
 
-interface PaymentInterface extends \Pimcore\Bundle\EcommerceFrameworkBundle\PaymentManager\Payment\PaymentInterface
+interface PaymentInterface
 {
+
+    /**
+     * @return string
+     */
+    public function getName();
+
     /**
      * Start payment
      *
@@ -44,4 +51,56 @@ interface PaymentInterface extends \Pimcore\Bundle\EcommerceFrameworkBundle\Paym
      * @return StartPaymentResponseInterface
      */
     public function startPayment(OrderAgentInterface $orderAgent, PriceInterface $price, AbstractRequest $config): StartPaymentResponseInterface;
+
+    /**
+     * Handles response of payment provider and creates payment status object
+     *
+     * @param StatusInterface $response
+     *
+     * @return StatusInterface
+     */
+    public function handleResponse($response);
+
+    /**
+     * Returns the authorized data from payment provider
+     *
+     * @return array
+     */
+    public function getAuthorizedData();
+
+    /**
+     * Set authorized data from payment provider
+     *
+     * @param array $authorizedData
+     */
+    public function setAuthorizedData(array $authorizedData);
+
+    /**
+     * Executes payment
+     *
+     * @param PriceInterface|null $price
+     * @param string|null $reference
+     *
+     * @return StatusInterface
+     */
+    public function executeDebit(PriceInterface $price = null, $reference = null);
+
+    /**
+     * Executes credit
+     *
+     * @param PriceInterface $price
+     * @param string $reference
+     * @param string $transactionId
+     *
+     * @return StatusInterface
+     */
+    public function executeCredit(PriceInterface $price, $reference, $transactionId);
+
+    /**
+     * returns configuration key in yml configuration file
+     *
+     * @return string
+     */
+    public function getConfigurationKey();
+
 }
