+++ bundles/EcommerceFrameworkBundle/Resources/config/checkout_manager.yml
@@ -22,12 +22,6 @@ services:
             - '@?'
             - '@event_dispatcher'
 
-    Pimcore\Bundle\EcommerceFrameworkBundle\CheckoutManager\CommitOrderProcessor:
-        arguments:
-            - '@Symfony\Component\Lock\LockFactory'
-            - '@pimcore_ecommerce.locator.order_manager'
-            - [] # options
-
     Pimcore\Bundle\EcommerceFrameworkBundle\CheckoutManager\V7\CommitOrderProcessor:
         arguments:
             - '@Symfony\Component\Lock\LockFactory'
