+++ bundles/EcommerceFrameworkBundle/Resources/config/order_manager.yml
@@ -11,18 +11,11 @@ services:
             - '@event_dispatcher'
             - []
 
-    Pimcore\Bundle\EcommerceFrameworkBundle\OrderManager\OrderManager:
-        arguments:
-            - '@pimcore_ecommerce.environment'
-            - '@?'
-            - '@pimcore_ecommerce.voucher_service'
-            - []
-
-
     Pimcore\Bundle\EcommerceFrameworkBundle\OrderManager\V7\OrderManager:
         arguments:
             - '@pimcore_ecommerce.environment'
             - '@?'
             - '@pimcore_ecommerce.voucher_service'
             - '@event_dispatcher'
+            - '@Pimcore\Model\FactoryInterface'
             - []
